order screen integrated

+++ routers/categories.js
@@ -23,6 +23,7 @@ router.get('/:id', async (req, res) => {
 
 router.post('/', async (req, res, next) => {
   if (!verifyAdmin(req, res)) {
+    console.log(2);
     let category = new Category({
       name: req.body.name,
       icon: req.body.icon,
@@ -36,36 +37,41 @@ router.post('/', async (req, res, next) => {
 });
 
 router.put('/:id', async (req, res) => {
-  let category = await Category.findByIdAndUpdate(
-    req.params.id,
-    {
-      name: req.body.name,
-      icon: req.body.icon,
-      color: req.body.color,
-    },
-    {
-      new: true,
-    }
-  );
-  if (!category) return res.status(500).send('The category can not be updated');
-  res.status(200).send(category);
+  if (!verifyAdmin(req, res)) {
+    let category = await Category.findByIdAndUpdate(
+      req.params.id,
+      {
+        name: req.body.name,
+        icon: req.body.icon,
+        color: req.body.color,
+      },
+      {
+        new: true,
+      }
+    );
+    if (!category)
+      return res.status(500).send('The category can not be updated');
+    res.status(200).send(category);
+  }
 });
 
 router.delete('/:id', (req, res) => {
-  Category.findByIdAndRemove(req.params.id)
-    .then((category) => {
-      if (category)
-        return res
-          .status(200)
-          .json({ success: true, message: 'The category is deleted!' });
-      else
-        return res
-          .status(404)
-          .json({ success: false, message: 'category not found' });
-    })
-    .catch((err) => {
-      return res.status(400).json({ success: 'false', error: err });
-    });
+  if (!verifyAdmin(req, res)) {
+    Category.findByIdAndRemove(req.params.id)
+      .then((category) => {
+        if (category)
+          return res
+            .status(200)
+            .json({ success: true, message: 'The category is deleted!' });
+        else
+          return res
+            .status(404)
+            .json({ success: false, message: 'category not found' });
+      })
+      .catch((err) => {
+        return res.status(400).json({ success: 'false', error: err });
+      });
+  }
 });
 
 module.exports = router;
